COMMERCE-5282 Extract dispatch logic into a helper, call update method on settings changes
