+++ modules/dxp/apps/commerce-avalara/commerce-avalara-connector-api/build.gradle
@@ -12,6 +12,7 @@ dependencies {
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
 	compileOnly project(":apps:commerce:commerce-account-api")
 	compileOnly project(":apps:commerce:commerce-api")
+	compileOnly project(":apps:commerce:commerce-tax-api")
 	compileOnly project(":apps:commerce:commerce-tax-engine-fixed-api")
 	compileOnly project(":core:petra:petra-string")
 }
