+++ modules/dxp/apps/commerce-avalara/commerce-avalara-connector/build.gradle
@@ -11,8 +11,10 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "osgi.core", version: "6.0.0"
 	compileOnly project(":apps:commerce:commerce-api")
+	compileOnly project(":apps:commerce:commerce-product-api")
 	compileOnly project(":apps:commerce:commerce-tax-api")
 	compileOnly project(":apps:commerce:commerce-tax-engine-fixed-api")
+	compileOnly project(":apps:dispatch:dispatch-api")
 	compileOnly project(":core:petra:petra-string")
 	compileOnly project(":dxp:apps:commerce-avalara:commerce-avalara-connector-api")
