+++ modules/dxp/apps/commerce-avalara/commerce-avalara-tax-engine-fixed-web/src/main/java/com/liferay/commerce/avalara/tax/engine/fixed/web/internal/portlet/action/EditCommerceTaxAvalaraMVCActionCommand.java
@@ -16,6 +16,7 @@ package com.liferay.commerce.avalara.tax.engine.fixed.web.internal.portlet.actio
 
 import com.liferay.commerce.avalara.connector.CommerceAvalaraConnector;
 import com.liferay.commerce.avalara.connector.configuration.CommerceAvalaraConnectorConfiguration;
+import com.liferay.commerce.avalara.connector.helper.CommerceAvalaraDispatchTriggerHelper;
 import com.liferay.commerce.constants.CommercePortletKeys;
 import com.liferay.commerce.tax.model.CommerceTaxMethod;
 import com.liferay.commerce.tax.service.CommerceTaxMethodService;
@@ -74,14 +75,21 @@ public class EditCommerceTaxAvalaraMVCActionCommand
 		}
 	}
 
-	private void _updateCommerceTaxAvalara(ActionRequest actionRequest)
+	private CommerceTaxMethod _getCommerceTaxMethod(ActionRequest actionRequest)
 		throws Exception {
 
 		long commerceTaxMethodId = ParamUtil.getLong(
 			actionRequest, "commerceTaxMethodId");
 
-		CommerceTaxMethod commerceTaxMethod =
-			_commerceTaxMethodService.getCommerceTaxMethod(commerceTaxMethodId);
+		return _commerceTaxMethodService.getCommerceTaxMethod(
+			commerceTaxMethodId);
+	}
+
+	private void _updateCommerceTaxAvalara(ActionRequest actionRequest)
+		throws Exception {
+
+		CommerceTaxMethod commerceTaxMethod = _getCommerceTaxMethod(
+			actionRequest);
 
 		Settings settings = _settingsFactory.getSettings(
 			new GroupServiceSettingsLocator(
@@ -116,6 +124,9 @@ public class EditCommerceTaxAvalaraMVCActionCommand
 		modifiableSettings.setValue("serviceURL", serviceURL);
 
 		modifiableSettings.store();
+
+		_commerceAvalaraDispatchTriggerHelper.updateDispatchTrigger(
+			_getCommerceTaxMethod(actionRequest));
 	}
 
 	private void _verifyConnection(ActionRequest actionRequest)
@@ -134,6 +145,10 @@ public class EditCommerceTaxAvalaraMVCActionCommand
 
 	@Reference
 	private CommerceAvalaraConnector _commerceAvalaraConnector;
+
+	@Reference
+	private CommerceAvalaraDispatchTriggerHelper
+		_commerceAvalaraDispatchTriggerHelper;
 
 	@Reference
 	private CommerceTaxMethodService _commerceTaxMethodService;
