+++ modules/dxp/apps/commerce-avalara/commerce-avalara-tax-engine-fixed/build.gradle
@@ -13,7 +13,6 @@ dependencies {
 	compileOnly project(":apps:commerce:commerce-product-api")
 	compileOnly project(":apps:commerce:commerce-tax-api")
 	compileOnly project(":apps:commerce:commerce-tax-engine-fixed-api")
-	compileOnly project(":apps:dispatch:dispatch-api")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib")
 	compileOnly project(":dxp:apps:commerce-avalara:commerce-avalara-connector-api")
 }
