+++ modules/dxp/apps/commerce-avalara/commerce-avalara-tax-engine-fixed/src/main/java/com/liferay/commerce/avalara/tax/engine/internal/model/listener/CommerceTaxMethodModelListener.java
@@ -15,21 +15,10 @@
 package com.liferay.commerce.avalara.tax.engine.internal.model.listener;
 
 import com.liferay.commerce.avalara.connector.constants.CommerceAvalaraConstants;
-import com.liferay.commerce.product.model.CommerceChannel;
-import com.liferay.commerce.product.service.CommerceChannelLocalService;
-import com.liferay.commerce.tax.engine.fixed.service.CommerceTaxFixedRateAddressRelLocalService;
-import com.liferay.commerce.tax.engine.fixed.service.CommerceTaxFixedRateLocalService;
+import com.liferay.commerce.avalara.connector.helper.CommerceAvalaraDispatchTriggerHelper;
 import com.liferay.commerce.tax.model.CommerceTaxMethod;
-import com.liferay.dispatch.model.DispatchTrigger;
-import com.liferay.dispatch.service.DispatchTriggerLocalService;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ModelListener;
-import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.UnicodeProperties;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -46,39 +35,8 @@ public class CommerceTaxMethodModelListener
 		String engineKey = commerceTaxMethod.getEngineKey();
 
 		if (engineKey.equals(CommerceAvalaraConstants.AVALARA)) {
-			try {
-				CommerceChannel commerceChannel = _getAssociatedCommerceChannel(
-					commerceTaxMethod);
-
-				String triggerName = _getAvalaraTriggerName(commerceChannel);
-
-				UnicodeProperties unicodeProperties = new UnicodeProperties(
-					true);
-
-				unicodeProperties.setProperty(
-					CommerceAvalaraConstants.GROUP_ID,
-					String.valueOf(commerceTaxMethod.getGroupId()));
-
-				DispatchTrigger dispatchTrigger =
-					_dispatchTriggerLocalService.addDispatchTrigger(
-						commerceTaxMethod.getUserId(), triggerName,
-						Boolean.FALSE, commerceTaxMethod.getEngineKey(),
-						unicodeProperties);
-
-				dispatchTrigger.setCronExpression(
-					_EVERY_MONTH_ON_THE_SECOND_CRON_EXPRESSION);
-
-				dispatchTrigger.setActive(Boolean.TRUE);
-
-				_dispatchTriggerLocalService.updateDispatchTrigger(
-					dispatchTrigger);
-			}
-			catch (PortalException portalException) {
-				_log.error(
-					"Could not create a dispatch trigger for newly created " +
-						"Avalara Tax Method",
-					portalException);
-			}
+			_commerceAvalaraDispatchTriggerHelper.createDispatchTrigger(
+				commerceTaxMethod);
 		}
 	}
 
@@ -87,26 +45,8 @@ public class CommerceTaxMethodModelListener
 		String engineKey = commerceTaxMethod.getEngineKey();
 
 		if (engineKey.equals(CommerceAvalaraConstants.AVALARA)) {
-			CommerceChannel commerceChannel = _getAssociatedCommerceChannel(
+			_commerceAvalaraDispatchTriggerHelper.updateDispatchTrigger(
 				commerceTaxMethod);
-
-			String triggerName = _getAvalaraTriggerName(commerceChannel);
-
-			DispatchTrigger dispatchTrigger =
-				_dispatchTriggerLocalService.fetchDispatchTrigger(
-					commerceChannel.getCompanyId(), triggerName);
-
-			if (dispatchTrigger != null) {
-				if (commerceTaxMethod.isActive()) {
-					dispatchTrigger.setActive(Boolean.TRUE);
-				}
-				else {
-					dispatchTrigger.setActive(Boolean.FALSE);
-				}
-
-				_dispatchTriggerLocalService.updateDispatchTrigger(
-					dispatchTrigger);
-			}
 		}
 	}
 
@@ -115,74 +55,13 @@ public class CommerceTaxMethodModelListener
 		String engineKey = commerceTaxMethod.getEngineKey();
 
 		if (engineKey.equals(CommerceAvalaraConstants.AVALARA)) {
-			try {
-				CommerceChannel commerceChannel = _getAssociatedCommerceChannel(
-					commerceTaxMethod);
-
-				String triggerName = _getAvalaraTriggerName(commerceChannel);
-
-				DispatchTrigger dispatchTrigger =
-					_dispatchTriggerLocalService.fetchDispatchTrigger(
-						commerceChannel.getCompanyId(), triggerName);
-
-				if (dispatchTrigger != null) {
-					_dispatchTriggerLocalService.deleteDispatchTrigger(
-						dispatchTrigger.getDispatchTriggerId());
-				}
-			}
-			catch (PortalException portalException) {
-				_log.error(
-					"Could not delete dispatch trigger for an Avalara Tax " +
-						"Method",
-					portalException);
-			}
+			_commerceAvalaraDispatchTriggerHelper.deleteDispatchTrigger(
+				commerceTaxMethod);
 		}
 	}
 
-	private CommerceChannel _getAssociatedCommerceChannel(
-		CommerceTaxMethod commerceTaxMethod) {
-
-		CommerceChannel commerceChannel = null;
-
-		try {
-			commerceChannel =
-				_commerceChannelLocalService.getCommerceChannelByGroupId(
-					commerceTaxMethod.getGroupId());
-		}
-		catch (PortalException portalException) {
-			_log.error("Could not get CommerceChannel", portalException);
-		}
-
-		return commerceChannel;
-	}
-
-	private String _getAvalaraTriggerName(CommerceChannel commerceChannel) {
-		StringBundler triggerNameSB = new StringBundler(3);
-
-		triggerNameSB.append(CommerceAvalaraConstants.AVALARA);
-		triggerNameSB.append(StringPool.DASH);
-		triggerNameSB.append(commerceChannel.getName());
-
-		return triggerNameSB.toString();
-	}
-
-	private static final String _EVERY_MONTH_ON_THE_SECOND_CRON_EXPRESSION =
-		"0 0 1 2 * ?";
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		CommerceTaxMethodModelListener.class);
-
 	@Reference
-	private CommerceChannelLocalService _commerceChannelLocalService;
-
-	@Reference
-	private CommerceTaxFixedRateAddressRelLocalService
-		_commerceTaxFixedRateAddressRelLocalService;
-
-	@Reference
-	private CommerceTaxFixedRateLocalService _commerceTaxFixedRateLocalService;
-
-	@Reference
-	private DispatchTriggerLocalService _dispatchTriggerLocalService;
+	private CommerceAvalaraDispatchTriggerHelper
+		_commerceAvalaraDispatchTriggerHelper;
 
 }
